move autocompletemanager from contact to utils
