+++ src/core/models/methods.ts
@@ -30,8 +30,8 @@ export class Methods {
     //Autocomplete handler
     static async autocompleteHandler(intera:AutocompleteInteraction) {
         let db = JSON.parse(readFileSync('./data/contacts.json', 'utf-8'));
-        let command = await import(`../commands/${intera.commandName.toLowerCase()}.js`);
-        await command.autocompleteManager(intera, db);
+        let command = await import(`../utils.js`);
+        await Utils.autocompleteManager(intera, db);
     }
 
     //Button handler
